Move filepicker shims into their own file
